Move remote sync stuff in FileProviderEnumerator to own file
